fairmetadata4j lib renaming issues fixed
