+++ src/main/java/nl/dtls/fairdatapoint/api/controller/MetadataController.java
@@ -55,16 +55,16 @@ import org.springframework.web.servlet.ModelAndView;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import nl.dtl.fairmetadata.io.MetadataException;
-import nl.dtl.fairmetadata.io.MetadataParserException;
-import nl.dtl.fairmetadata.model.Agent;
-import nl.dtl.fairmetadata.model.CatalogMetadata;
-import nl.dtl.fairmetadata.model.DatasetMetadata;
-import nl.dtl.fairmetadata.model.DistributionMetadata;
-import nl.dtl.fairmetadata.model.FDPMetadata;
-import nl.dtl.fairmetadata.model.Identifier;
-import nl.dtl.fairmetadata.utils.MetadataUtils;
-import nl.dtl.fairmetadata.utils.vocabulary.DataCite;
+import nl.dtl.fairmetadata4j.io.MetadataException;
+import nl.dtl.fairmetadata4j.io.MetadataParserException;
+import nl.dtl.fairmetadata4j.model.Agent;
+import nl.dtl.fairmetadata4j.model.CatalogMetadata;
+import nl.dtl.fairmetadata4j.model.DatasetMetadata;
+import nl.dtl.fairmetadata4j.model.DistributionMetadata;
+import nl.dtl.fairmetadata4j.model.FDPMetadata;
+import nl.dtl.fairmetadata4j.model.Identifier;
+import nl.dtl.fairmetadata4j.utils.MetadataUtils;
+import nl.dtl.fairmetadata4j.utils.vocabulary.DataCite;
 import nl.dtls.fairdatapoint.api.controller.utils.LoggerUtils;
 import nl.dtls.fairdatapoint.service.FairMetaDataService;
 import nl.dtls.fairdatapoint.service.FairMetadataServiceException;
