+++ src/main/java/nl/dtls/fairdatapoint/api/converter/DatasetMetadataConverter.java
@@ -33,12 +33,12 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.nio.charset.StandardCharsets;
-import nl.dtl.fairmetadata.io.DatasetMetadataParser;
-import nl.dtl.fairmetadata.io.MetadataException;
-import nl.dtl.fairmetadata.io.MetadataParserException;
-import nl.dtl.fairmetadata.model.DatasetMetadata;
-import nl.dtl.fairmetadata.utils.MetadataParserUtils;
-import nl.dtl.fairmetadata.utils.MetadataUtils;
+import nl.dtl.fairmetadata4j.io.DatasetMetadataParser;
+import nl.dtl.fairmetadata4j.io.MetadataException;
+import nl.dtl.fairmetadata4j.io.MetadataParserException;
+import nl.dtl.fairmetadata4j.model.DatasetMetadata;
+import nl.dtl.fairmetadata4j.utils.MetadataParserUtils;
+import nl.dtl.fairmetadata4j.utils.MetadataUtils;
 import org.eclipse.rdf4j.rio.RDFFormat;
 import org.springframework.http.HttpInputMessage;
 import org.springframework.http.HttpOutputMessage;
