+++ src/main/java/nl/dtls/fairdatapoint/service/impl/FairMetaDataServiceImpl.java
@@ -32,22 +32,22 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.annotation.Nonnull;
 import javax.xml.datatype.DatatypeConfigurationException;
-import nl.dtl.fairmetadata.io.CatalogMetadataParser;
-import nl.dtl.fairmetadata.io.DatasetMetadataParser;
-import nl.dtl.fairmetadata.io.DistributionMetadataParser;
-import nl.dtl.fairmetadata.io.FDPMetadataParser;
-import nl.dtl.fairmetadata.io.MetadataException;
-import nl.dtl.fairmetadata.model.CatalogMetadata;
-import nl.dtl.fairmetadata.model.DatasetMetadata;
-import nl.dtl.fairmetadata.model.DistributionMetadata;
-import nl.dtl.fairmetadata.model.FDPMetadata;
-import nl.dtl.fairmetadata.model.Metadata;
-import nl.dtl.fairmetadata.utils.MetadataParserUtils;
-import nl.dtl.fairmetadata.utils.MetadataUtils;
-import nl.dtl.fairmetadata.utils.RDFUtils;
-import nl.dtl.fairmetadata.utils.vocabulary.DCAT;
-import nl.dtl.fairmetadata.utils.vocabulary.FDP;
-import nl.dtl.fairmetadata.utils.vocabulary.R3D;
+import nl.dtl.fairmetadata4j.io.CatalogMetadataParser;
+import nl.dtl.fairmetadata4j.io.DatasetMetadataParser;
+import nl.dtl.fairmetadata4j.io.DistributionMetadataParser;
+import nl.dtl.fairmetadata4j.io.FDPMetadataParser;
+import nl.dtl.fairmetadata4j.io.MetadataException;
+import nl.dtl.fairmetadata4j.model.CatalogMetadata;
+import nl.dtl.fairmetadata4j.model.DatasetMetadata;
+import nl.dtl.fairmetadata4j.model.DistributionMetadata;
+import nl.dtl.fairmetadata4j.model.FDPMetadata;
+import nl.dtl.fairmetadata4j.model.Metadata;
+import nl.dtl.fairmetadata4j.utils.MetadataParserUtils;
+import nl.dtl.fairmetadata4j.utils.MetadataUtils;
+import nl.dtl.fairmetadata4j.utils.RDFUtils;
+import nl.dtl.fairmetadata4j.utils.vocabulary.DCAT;
+import nl.dtl.fairmetadata4j.utils.vocabulary.FDP;
+import nl.dtl.fairmetadata4j.utils.vocabulary.R3D;
 import nl.dtls.fairdatapoint.repository.StoreManager;
 import nl.dtls.fairdatapoint.repository.StoreManagerException;
 import nl.dtls.fairdatapoint.service.FairMetaDataService;
